+++ examples/form-tutorial/form_tutorial_4.js
@@ -6,7 +6,7 @@ var formConfig = {
     title: 'Form: Field Types',
     type : 'pong-form',
     resourceURL: 'hello',
-    height: '700px'
+    height: '750px'
   }
   
 var formPlugInConfig = {
@@ -28,13 +28,18 @@ var formPlugInConfig = {
               { id: 'f04', label: 'E-Mail', type: 'email' },
               { id: 'f05', label: 'Date', type: 'date' },
               { id: 'f06', label: 'Select', type: 'select',
-                options:[ {option:'first',value:'Option1'}, {option:'second',value:'Option2'}, {option:'third',value:'Option3'} ] },
+                options:[ {option:'first',value:'Option1'}, {option:'second',value:'Option2'}, {option:'third',value:'Option3'} ] }
+            ]
+          },
+          {
+            formFields: [
+              { id: 'f00', label: 'Enable POST', type: 'checkbox', name:'accept', activate:["actionPost"] },
               { id: 'fs1', label: 'Separator', type: 'separator' },
-              { id: 'f07a', label: 'This is a "label"', type: 'label' },
-              { id: 'f07a', label: 'Checkbox 1st', type: 'checkbox', name:'myChecks', value:'1', defaultVal:true, descr:'is read-only', readonly:true },
-              { id: 'f07b', label: 'Checkbox 2nd', type: 'checkbox', name:'myChecks', value:'2', defaultVal:'true' },
-              { id: 'f07b', label: 'Checkbox 2nd', type: 'checkbox', name:'myChecks', value:'3', defaultVal:false },
+              { id: 'f07a', label: 'RO-Checkbox', type: 'checkbox', name:'myChecks', value:'1', defaultVal:true, descr:'is read-only', readonly:true },
+              { id: 'f07b', label: '2nd Checkbox', type: 'checkbox', name:'myChecks', value:'2', defaultVal:'true' },
+              { id: 'f07c', label: '3rd Checkbox', type: 'checkbox', name:'myChecks', value:'3', defaultVal:false },
               { id: 'fs2', label: 'Separator', type: 'separator' },
+              { id: 'f07', label: 'This is a "label"', type: 'label' },
               { id: 'f08a', label: 'Radio 1st', type: 'radio', name:'myRadio', value:'1st' },
               { id: 'f08b', label: 'Radio 2nd', type: 'radio', name:'myRadio', value:'2nd' }
             ]
@@ -44,7 +49,7 @@ var formPlugInConfig = {
     ],
     actions: [
       {  id: 'actionGet',  actionName: 'GET',  method: 'GET',  actionURL: 'test' },
-      {  id: 'actionPost', actionName: 'POST', method: 'POST', actionURL: 'test' }      
+      {  id: 'actionPost', actionName: 'POST', method: 'POST', actionURL: 'test', enabled:false }      
     ]
   }