+++ ReleaseNotes.md
@@ -6,6 +6,7 @@
 - plan HTML fix and added to [complex layout example](https://github.com/ma-ha/easy-web-app/tree/master/examples/complex-layout)
 - fix ignored page layouts in nav
 - table: fix editable checkbox 
+- form checkbox: enable/disable other form elements
 
 ## 0.17.x
 - Table/List: support number